requirements.txt
Django
djangorestframework
requests 